% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project_template.R
\name{make_project_template}
\alias{make_project_template}
\title{Make Project Template}
\usage{
make_project_template(path, ...)
}
\arguments{
\item{path}{The path to the new project.}
}
\description{
Create a new project template in RStudio.
}
\examples{
try(make_project_template("path/to/new/project"))
}
